Move do_setcontext from ucontext.c to context.c
