--- sys/aarch64/ucontext.c
@@ -1,7 +1,0 @@
-#include <sys/mimiker.h>
-#include <sys/thread.h>
-#include <aarch64/mcontext.h>
-
-int do_setcontext(thread_t *td, ucontext_t *uc) {
-  panic("Not implemented!");
-}
