+++ sys/mips/Makefile
@@ -22,8 +22,7 @@ SOURCES = \
 	timer.c \
 	tlb.c \
 	trap.c \
-	uart_cbus.c \
-	ucontext.c
+	uart_cbus.c
 
 OBJECTS += $(BOARD)_dtb.o
